Make sure to cast SIZE_MAX to size_t in tests.
Otherwise, macOS complains.

+++ test/test_find.cpp
@@ -57,13 +57,18 @@ size_t test_find_lastn(const char * str, char delimiter, size_t str_len, size_t 
 
 TEST(test_find, find) {
   size_t ret0 = test_find("", '/', SIZE_MAX);
-  LOG(SIZE_MAX, ret0);
+  // We cast SIZE_MAX to a size_t here (and below) to shut up warnings on macOS.
+  // The problem on macOS is that size_t is a long unsigned int, while SIZE_MAX
+  // is an unsigned long long.  While the two are compatible on 64-bit, they are
+  // not considered the "same" by the compiler, and throw a warning.  Just cast
+  // SIZE_MAX to size_t to make the compiler happy.
+  LOG((size_t)SIZE_MAX, ret0);
 
   size_t ret00 = test_find(NULL, '/', SIZE_MAX);
-  LOG(SIZE_MAX, ret00);
+  LOG((size_t)SIZE_MAX, ret00);
 
   size_t ret1 = test_find("hello_world", '/', SIZE_MAX);
-  LOG(SIZE_MAX, ret1);
+  LOG((size_t)SIZE_MAX, ret1);
 
   size_t ret2 = test_find("hello/world", '/', 5);
   LOG((size_t)5, ret2);
@@ -83,13 +88,13 @@ TEST(test_find, find) {
 
 TEST(test_find, findn) {
   size_t ret0 = test_findn("", '/', 0, SIZE_MAX);
-  LOG(SIZE_MAX, ret0);
+  LOG((size_t)SIZE_MAX, ret0);
 
   size_t ret1 = test_findn(NULL, '/', 10, SIZE_MAX);
-  LOG(SIZE_MAX, ret1);
+  LOG((size_t)SIZE_MAX, ret1);
 
   size_t ret2 = test_findn("hello_world", '/', strlen("hello_world"), SIZE_MAX);
-  LOG(SIZE_MAX, ret2);
+  LOG((size_t)SIZE_MAX, ret2);
 
   size_t ret3 = test_findn("hello/world", '/', strlen("hello/world"), 5);
   LOG((size_t)5, ret3);
@@ -100,13 +105,13 @@ TEST(test_find, findn) {
 
 TEST(test_find, find_last) {
   size_t ret0 = test_find_last("", '/', SIZE_MAX);
-  LOG(SIZE_MAX, ret0);
+  LOG((size_t)SIZE_MAX, ret0);
 
   size_t ret00 = test_find_last(NULL, '/', SIZE_MAX);
-  LOG(SIZE_MAX, ret00);
+  LOG((size_t)SIZE_MAX, ret00);
 
   size_t ret1 = test_find_last("hello_world", '/', SIZE_MAX);
-  LOG(SIZE_MAX, ret1);
+  LOG((size_t)SIZE_MAX, ret1);
 
   size_t ret2 = test_find_last("hello/world", '/', 5);
   LOG((size_t)5, ret2);
@@ -126,13 +131,13 @@ TEST(test_find, find_last) {
 
 TEST(test_find, find_lastn) {
   size_t ret0 = test_find_lastn("", '/', 0, SIZE_MAX);
-  LOG(SIZE_MAX, ret0);
+  LOG((size_t)SIZE_MAX, ret0);
 
   size_t ret1 = test_find_lastn(NULL, '/', 10, SIZE_MAX);
-  LOG(SIZE_MAX, ret1);
+  LOG((size_t)SIZE_MAX, ret1);
 
   size_t ret2 = test_find_lastn("hello_world", '/', strlen("hello_world"), SIZE_MAX);
-  LOG(SIZE_MAX, ret2);
+  LOG((size_t)SIZE_MAX, ret2);
 
   size_t ret3 = test_find_lastn("hello/world", '/', strlen("hello/world"), 5);
   LOG((size_t)5, ret3);
